Instructions use bundle add to reduce steps
If you use `bundle add` here you get to take out the instructions to edit the Gemfile and the command to bundle afterward, it does it all in one step. You might want to check on when they added the `add` command to make sure it has the version coverage you'd like but I'm willing to bet that your customer base will be using a modern version most of the time.